LPS-56065 Move MVC actions into module
