+++ modules/apps/portal-settings/portal-settings-web/src/com/liferay/portalsettings/web/action/EditCompanyMVCActionCommand.java
@@ -12,7 +12,14 @@
  * details.
  */
 
-package com.liferay.portlet.portalsettings.action;
+package com.liferay.portalsettings.web.action;
+
+import java.util.List;
+
+import javax.portlet.ActionRequest;
+import javax.portlet.ActionResponse;
+
+import org.osgi.service.component.annotations.Component;
 
 import com.liferay.portal.AccountNameException;
 import com.liferay.portal.AddressCityException;
@@ -32,7 +39,6 @@ import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.servlet.SessionErrors;
-import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
 import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.FileUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
@@ -52,17 +58,12 @@ import com.liferay.portal.util.PropsValues;
 import com.liferay.portlet.documentlibrary.service.DLAppLocalServiceUtil;
 import com.liferay.portlet.usersadmin.util.UsersAdminUtil;
 
-import java.util.List;
-
-import javax.portlet.ActionRequest;
-import javax.portlet.ActionResponse;
-
 /**
  * @author Brian Wing Shun Chan
  * @author Julio Camarero
  * @author Philip Jones
  */
-@OSGiBeanProperties(
+@Component(
 	property = {
 		"javax.portlet.name=" + PortletKeys.PORTAL_SETTINGS,
 		"mvc.command.name=/portal_settings/edit_company"
