+++ modules/apps/portal-settings/portal-settings-web/src/com/liferay/portalsettings/web/action/EditLDAPServerMVCRenderCommand.java
@@ -12,31 +12,30 @@
  * details.
  */
 
-package com.liferay.portlet.portalsettings.action;
-
-import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
-import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
-import com.liferay.portal.util.PortletKeys;
+package com.liferay.portalsettings.web.action;
 
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
 
+import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
+import com.liferay.portal.util.PortletKeys;
+
 /**
  * @author Philip Jones
  */
-@OSGiBeanProperties(
+@Compenent(
 	property = {
 		"javax.portlet.name=" + PortletKeys.PORTAL_SETTINGS,
-		"mvc.command.name=/portal_settings/test_ldap_users"
+		"mvc.command.name=/portal_settings/edit_ldap_server"
 	}
 )
-public class TestLDAPUsersMVCRenderCommand implements MVCRenderCommand {
+public class EditLDAPServerMVCRenderCommand implements MVCRenderCommand {
 
 	@Override
 	public String render(
 		RenderRequest renderRequest, RenderResponse renderResponse) {
 
-		return "/html/portlet/portal_settings/test_ldap_users.jsp";
+		return "/html/portlet/portal_settings/edit_ldap_server.jsp";
 	}
 
 }
