+++ modules/apps/portal-settings/portal-settings-web/src/com/liferay/portalsettings/web/action/TestLDAPUsersMVCRenderCommand.java
@@ -12,32 +12,32 @@
  * details.
  */
 
-package com.liferay.portlet.portalsettings.action;
-
-import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
-import com.liferay.portal.kernel.spring.osgi.OSGiBeanProperties;
-import com.liferay.portal.util.PortletKeys;
+package com.liferay.portalsettings.web.action;
 
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
 
+import org.osgi.service.component.annotations.Component;
+
+import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
+import com.liferay.portal.util.PortletKeys;
+
 /**
  * @author Philip Jones
  */
-@OSGiBeanProperties(
+@Component(
 	property = {
 		"javax.portlet.name=" + PortletKeys.PORTAL_SETTINGS,
-		"mvc.command.name=/portal_settings/test_opensso_configuration"
+		"mvc.command.name=/portal_settings/test_ldap_users"
 	}
 )
-public class TestOpenSSOConfigurationMVCRenderCommand
-	implements MVCRenderCommand {
+public class TestLDAPUsersMVCRenderCommand implements MVCRenderCommand {
 
 	@Override
 	public String render(
 		RenderRequest renderRequest, RenderResponse renderResponse) {
 
-		return "/html/portlet/portal_settings/test_opensso_configuration.jsp";
+		return "/html/portlet/portal_settings/test_ldap_users.jsp";
 	}
 
 }
