+++ portal-impl/src/META-INF/util-spring.xml
@@ -598,15 +598,6 @@
 	<bean id="com.liferay.portlet.login.action.ForgotPasswordMVCRenderCommand" class="com.liferay.portlet.login.action.ForgotPasswordMVCRenderCommand" />
 	<bean id="com.liferay.portlet.login.action.LoginMVCActionCommand" class="com.liferay.portlet.login.action.LoginMVCActionCommand" />
 	<bean id="com.liferay.portlet.login.action.LoginMVCRenderCommand" class="com.liferay.portlet.login.action.LoginMVCRenderCommand" />
-	<bean id="com.liferay.portlet.portalsettings.action.EditCompanyMVCActionCommand" class="com.liferay.portlet.portalsettings.action.EditCompanyMVCActionCommand" />
-	<bean id="com.liferay.portlet.portalsettings.action.EditCompanyMVCRenderCommand" class="com.liferay.portlet.portalsettings.action.EditCompanyMVCRenderCommand" />
-	<bean id="com.liferay.portlet.portalsettings.action.EditLDAPServerMVCActionCommand" class="com.liferay.portlet.portalsettings.action.EditLDAPServerMVCActionCommand" />
-	<bean id="com.liferay.portlet.portalsettings.action.EditLDAPServerMVCRenderCommand" class="com.liferay.portlet.portalsettings.action.EditLDAPServerMVCRenderCommand" />
-	<bean id="com.liferay.portlet.portalsettings.action.TestCASConfigurationMVCRenderCommand" class="com.liferay.portlet.portalsettings.action.TestCASConfigurationMVCRenderCommand" />
-	<bean id="com.liferay.portlet.portalsettings.action.TestLDAPConnectionMVCRenderCommand" class="com.liferay.portlet.portalsettings.action.TestLDAPConnectionMVCRenderCommand" />
-	<bean id="com.liferay.portlet.portalsettings.action.TestLDAPGroupsMVCRenderCommand" class="com.liferay.portlet.portalsettings.action.TestLDAPGroupsMVCRenderCommand" />
-	<bean id="com.liferay.portlet.portalsettings.action.TestLDAPUsersMVCRenderCommand" class="com.liferay.portlet.portalsettings.action.TestLDAPUsersMVCRenderCommand" />
-	<bean id="com.liferay.portlet.portalsettings.action.TestOpenSSOConfigurationMVCRenderCommand" class="com.liferay.portlet.portalsettings.action.TestOpenSSOConfigurationMVCRenderCommand" />
 	<bean id="com.liferay.portlet.rolesadmin.util.RolesAdmin" class="com.liferay.portlet.rolesadmin.util.RolesAdminImpl" />
 	<bean id="com.liferay.portlet.rolesadmin.util.RolesAdminUtil" class="com.liferay.portlet.rolesadmin.util.RolesAdminUtil">
 		<property name="rolesAdmin" ref="com.liferay.portlet.rolesadmin.util.RolesAdmin" />
